Benchmarks
Google Benchmark added as a dependency and a separate executable is created for benchmarks.
A basic benchmark is added for inserting X and finding 10X times randomly.
insert function is added to skip list for when the inserted iterator is required, or to make using algorithms easier.
a flip function is added which provides much faster random boolean generation. this problem was pointed by Intel VTune profiler.